email_agent/requirements.txt
msal>=1.28
requests>=2.31
pyyaml>=6.0
fastapi>=0.111
uvicorn[standard]>=0.30
apscheduler>=3.10
anthropic>=0.28
